SRC/requirements.txt
Django==4.2.15
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
reportlab==4.2.2